Refactor redrawing methods and global variables
