+++ src/index.html
@@ -34,12 +34,12 @@
         </div>
 
       </div class="grid-element searchbar">
-          <div class="autocomplete">
-            <input id="searchbar" name="searchbar" type="text" class="searchbar" placeholder="Search"/>
-          </div>
+        <div class="autocomplete">
+          <input id="searchbar" name="searchbar" type="text" class="searchbar" placeholder="Search"/>
         </div>
-
       </div>
-      <script src="./index.js"></script>
-    </body>
-    </html>
+
+    </div>
+    <script src="./index.js"></script>
+  </body>
+</html>
