+++ src/index.js
@@ -12,261 +12,74 @@ import {
 import {
   medalslider
 } from './medalslider';
+var {
+  outerWidth,
+  outerHeight,
+  margin,
+  innerWidth,
+  innerHeight,
+  circleRadius,
+  xColumn,
+  yColumn,
+  colorColumn,
+  startYear,
+  endYear,
 
-var outerWidth = 1200;
-var outerHeight = 800;
-var margin = {
-  left: 60,
-  top: 30,
-  right: 30,
-  bottom: 30
-};
-var innerWidth = outerWidth - margin.left - margin.right;
-var innerHeight = outerHeight - margin.top - margin.bottom;
-var circleRadius = 3;
-var xColumn = "Order";
-var yColumn = "Year";
-var colorColumn = "NOC"; // color of circles based on athlete NOC
-var startYear = 1896;
-var endYear = 2016;
+  minOrder,
+  maxOrder, 
+  // csvFile,
 
-var minOrder = 0;
-var maxOrder = 8760;
+  medalCounts,
+  medalRange,
+  selectedValues,
+  peopleNames,
+  // data structures to be loaded in
+
+  // make the Name the key
+  entriesByName,
+  // make the NOC the key
+  entriesByNOC,
+  // make the Start Year the key then Name the secondary key
+  entriesByStartThenName,
+  // lists all NOCs
+  NOCs,
+
+  xScale,
+  yScale, 
+  colorScale, 
+  // get the svg
+  svg,
+
+  // for plotting points
+  xValue,
+
+  // axes
+  xAxis,
+  yAxis,
+
+  // add axis groups to svg
+  xAxisGroup,
+  yAxisGroup,
+
+  // define clipping path so that drawing stays within the chart
+  // while using the timeslider
+  clippath,
+
+  // add group for the visualization
+  chart
+} = require('./globals');
+import {
+  redraw,
+  redrawWithAnimation
+} from './canvas';
 
 const csvFile = require('../data/olympic_overall.csv');
-
-var medalCounts;
-var medalRange = [1, 28]; // [minMedals, maxMedals]
-var selectedValues = [];
-var peopleNames = [];
-// data structures to be loaded in
-
-// make the Name the key
-var entriesByName;
-// make the NOC the key
-var entriesByNOC;
-// make the Start Year the key then Name the secondary key
-var entriesByStartThenName;
-// lists all NOCs
-var NOCs = [];
-
-
-///////////////////////////////////////////////////////
-///////////////////////////////////////////////////////
-// HTML SETUP
-///////////////////////////////////////////////////////
-///////////////////////////////////////////////////////
-
-
-// const xScale = d3.scaleLinear().domain([minID, maxID]).range([margin["left"], innerWidth]);
-const xScale = d3.scaleLinear().domain([minOrder, maxOrder]).range([margin["left"], innerWidth]);
-
-//const xScale = d3.scalePoint().range([margin["left"], innerWidth]);
-const yScale = d3.scaleTime().domain([startYear, endYear]).range([margin["bottom"], innerHeight]);
-const colorScale = d3.scaleOrdinal(d3.schemeCategory10);
-
-// get the svg
-var svg = d3.select('svg');
-
-// for plotting points
-const xValue = d => d.Order;
-
-// axes
-//var xAxis = d3.axisBottom(xScale);
-var xAxis = d3.axisBottom(xScale)
-  .tickPadding(30)
-  .tickSize(0);
-var yAxis = d3.axisLeft(yScale)
-  .tickValues([1896, 1900, 1904, 1908, 1912, 1916, 1920, 1924, 1928, 1932,
-    1936, 1940, 1944, 1948, 1952, 1956, 1960, 1964, 1968, 1972, 1976, 1980,
-    1984, 1988, 1992, 1996, 2000, 2004, 2008, 2012, 2016
-  ])
-  .tickFormat(d3.format("Y")) // gets rid of commas in the dates
-  .tickSize(-innerWidth);
-
-// add axis groups to svg
-var xAxisGroup = svg.append("g")
-  .attr("class", "axis x")
-  .attr("transform", "translate(0," + innerHeight + ")")
-  .call(xAxis);
-var yAxisGroup = svg.append("g")
-  .attr("class", "axis y")
-  .attr("transform", "translate(" + margin["left"] + ",0)")
-  .call(yAxis);
-
-// define clipping path so that drawing stays within the chart
-// while using the timeslider
-var clippath = svg.append("clipPath")
-  .attr("id", "chart-clip")
-  .append("rect")
-  .attr("x", 0)
-  .attr("y", margin["bottom"] - 6) // magic number :o
-  .attr("width", outerWidth)
-  .attr("height", yAxisGroup.node().getBBox().height); //innerHeight
-
-// apply clipping path
-svg.attr("clip-path", "url(#chart-clip)");
-
-// add group for the visualization
-var chart = svg.append("g")
-  .attr("class", "chart")
-  .attr("transform", "translate(10,0)");
 
 ///////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////
 // Helper functions
 ///////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////
-
-// noc is the one you wanna remove
-
-// function to draw lines and points given inputData
-// Note: currently does not support .exit() functionality
-function redraw(inputData) {
-  if (typeof inputData !== 'undefined') {
-
-    // update the X-axis
-    xScale.domain([minOrder, maxOrder]);
-    xAxisGroup.transition().call(xAxis);
-
-    var line = chart.append("g").selectAll("line").data(inputData.values)
-      .enter()
-      .append("line")
-      .style("opacity", 0)
-      .style("stroke", function(d) {
-        return colorScale(d[colorColumn]);
-      })
-      .style("stroke-width", 1)
-      .attr("id", function(d) {
-        return "l" + d.NOC;
-      })
-      .attr("x1", function(d) {
-        return xScale(d[xColumn]);
-      })
-      .attr("y1", function(d) {
-        return yScale(d["Start"]);
-      })
-      .attr("x2", function(d) {
-        return xScale(d[xColumn]);
-      })
-      .attr("y2", function(d) {
-        return yScale(d["End"]);
-      });
-
-      line.transition()
-      .delay(3500)
-      .ease(d3.easeLinear)
-      .duration(1250)
-      .style("opacity", 1.0)
-
-    var circle = chart.append("g").selectAll('circle').data(inputData.values)
-      .enter()
-      .append('circle')
-      .style("opacity", 0)
-      .attr("id", function(d) {
-        return "c" + d.NOC;
-      })
-      .attr("cx", function(d) {
-        // return xScale(d[xColumn]);
-        return xScale((maxOrder + minOrder) / 2);
-      })
-      .attr("cy", function(d) {
-        return yScale(d[yColumn]);
-      })
-      .attr("r", circleRadius)
-      .attr("fill", "grey")
-      .attr("label", function(d) {
-        return d.Name
-      })
-
-    circle.transition()
-      .delay(1000)
-      .duration(1000)
-      .style("opacity", 1.0)
-      .attr("cx", function(d) {
-        // return xScale(d[xColumn]);
-        return xScale((maxOrder + minOrder) / 2);
-      })
-      .attr("cy", function(d) {
-        return yScale(d[yColumn]);
-      })
-      .attr("r", circleRadius)
-      .attr("fill", "grey")
-      .attr("label", function(d) {
-        return d.Name
-      })
-
-    circle.transition()
-      .delay(1500)
-      .duration(2000)
-      .ease(d3.easeQuadIn)
-      .attr("cx", function(d) {
-        return xScale(d[xColumn]);
-      })
-      .attr("cy", function(d) {
-        return yScale(d[yColumn]);
-      })
-      .attr("fill", function(d) {
-        return colorScale(d[colorColumn]);
-      })
-
-    circle.on("mouseover", function(d) {
-        // circle gets bigger
-        d3.select(this)
-          .transition()
-          .attr("r", circleRadius + 3)
-          .attr("fill", "orange");
-        //Get this circle's x/y values, then augment for the tooltip
-        var xPosition = parseFloat(d3.select(this).attr("cx"));
-        var yPosition = parseFloat(d3.select(this).attr("cy"));
-        console.log("x: " + xPosition + " y: " + yPosition);
-        //Create the tooltip label
-        svg.append("text")
-          .attr("id", "tooltip")
-          .attr("x", xPosition)
-          .attr("y", yPosition)
-          .attr("text-anchor", "right")
-          .attr("font-family", "sans-serif")
-          .attr("font-size", "11px")
-          .attr("font-weight", "bold")
-          .attr("fill", "black")
-          .style("pointer-events", "none")
-          .text(d.Name);
-
-        // get the data for the selected athlete
-        var athleteData = _.find(d3.values(entriesByName), function(item) {
-          return item.key === d.Name;
-        });
-        generateAthleteChart(athleteData.values);
-      })
-      .on("mouseout", function() {
-        // back to small circles
-        d3.select(this)
-          .transition()
-          .attr("r", circleRadius)
-          .attr("fill", function(d) {
-            return colorScale(d[colorColumn]);
-          });
-        // Remove the tooltip
-        d3.select("#tooltip").remove();
-      })
-      .on("click", function(d) {
-        // get the data for the selected athlete
-        console.log(Object.values(d3.values(entriesByName)));
-        var athleteData = _.find(d3.values(entriesByName), function(item) {
-          // console.log("key: " + item.key);
-          // console.log("searching for: " + d.Name);
-          return item.key === d.Name;
-        });
-        console.log("result: " + athleteData.values);
-
-        athleteData.values.forEach(function(element) {
-          console.log(element);
-        });
-
-      });
-  }
-}
 
 // function to set up the nested data structures
 function initializeDataStructures(data) {
@@ -390,7 +203,8 @@ function setupNOCFiltering(data) {
     }
     peopleNames = [];
     for (var i = 0; i < selectedValues.length; i++) {
-      redraw(filterByMedal(entriesByNOC[selectedValues[i]], medalCounts, medalRange[0], medalRange[1]));
+      redrawWithAnimation(svg, chart, filterByMedal(entriesByNOC[selectedValues[i]], medalCounts, medalRange[0], medalRange[1]),
+         entriesByName, xScale, yScale, colorScale, xAxis, xAxisGroup, xColumn, yColumn, colorColumn, circleRadius, minOrder, maxOrder);
     }
 
   });
@@ -441,7 +255,8 @@ function initializeMedalSlider() {
       // redraw data within range selection
       if (typeof selectedValues !== 'undefined') {
         for (let i = 0; i < selectedValues.length; i++) {
-          redraw(filterByMedal(entriesByNOC[selectedValues[i]], medalCounts, medalRange[0], medalRange[1]));
+          redraw(svg, chart, filterByMedal(entriesByNOC[selectedValues[i]], medalCounts, medalRange[0], medalRange[1]),
+             entriesByName, xScale, yScale, colorScale, xAxis, xAxisGroup, xColumn, yColumn, colorColumn, circleRadius, minOrder, maxOrder);
         }
       }
     });
